+++ inst/doc/supportR.R
@@ -5,6 +5,7 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 # devtools::install_github("njlyon0/supportR")
 
 ## ----setup--------------------------------------------------------------------
+#install.packages("supportR")
 library(supportR)
 
 ## ----ex-data-pngn-------------------------------------------------------------
@@ -12,12 +13,42 @@ library(supportR)
 library(palmerpenguins)
 
 # Glimpse the penguins dataset
-dplyr::glimpse(penguins)
+str(penguins)
 
 ## ----summary_table------------------------------------------------------------
 # Summarize the data
-summary_table(data = penguins, groups = c("species", "island"),
-              response = "bill_length_mm", drop_na = T)
+supportR::summary_table(data = penguins, groups = c("species", "island"),
+                        response = "bill_length_mm", drop_na = T)
+
+## ----crop_tri-----------------------------------------------------------------
+# Define a simple matrix wtih symmetric dimensions
+mat <- matrix(data = c(1:2, 2:1), nrow = 2, ncol = 2)
+
+# Crop off it's lower triangle
+supportR::crop_tri(data = mat, drop_tri = "lower", drop_diag = FALSE)
+
+# Drop the diagonal as well
+supportR::crop_tri(data = mat, drop_tri = "lower", drop_diag = TRUE)
+
+## ----array_melt---------------------------------------------------------------
+# Make data to fill the array
+vec1 <- c(5, 9, 3)
+vec2 <- c(10:15)
+
+# Create dimension names (x = col, y = row, z = which matrix)
+x_vals <- c("Col_1","Col_2","Col_3")
+y_vals <- c("Row_1","Row_2","Row_3")
+z_vals <- c("Mat_1","Mat_2")
+
+# Make an array from these components
+g <- array(data = c(vec1, vec2), dim = c(3, 3, 2),
+           dimnames = list(x_vals, y_vals, z_vals))
+
+# "Melt" the array into a dataframe
+melted <- supportR::array_melt(array = g)
+
+# Look at that top of that
+head(melted)
 
 ## ----diff_check---------------------------------------------------------------
 # Make two vectors
@@ -25,7 +56,7 @@ vec1 <- c("x", "a", "b")
 vec2 <- c("y", "z", "a")
 
 # Compare them!
-diff_check(old = vec1, new = vec2)
+supportR::diff_check(old = vec1, new = vec2, sort = TRUE, return = TRUE)
 
 ## ----num_check----------------------------------------------------------------
 # Make a dataframe with non-numbers in a number column
@@ -41,7 +72,7 @@ sites <- data.frame('site' = c("LTR", "GIL", "PYN", "RIN"),
                     'visit' = c('2021-01-01', '2021-01-0w', '1990', '2020-10-xx'))
 
 # Now we can use our function to identify bad dates
-date_check(data = sites, col = 'visit')
+supportR::date_check(data = sites, col = 'visit')
 
 ## ----date_format_guess--------------------------------------------------------
 # Make a dataframe with dates in various formats and a grouping column
@@ -57,23 +88,27 @@ my_df <- data.frame('data_enterer' = c('person A', 'person B',
                                     '6/10/02'))
 
 # Now we can invoke the function!
-date_format_guess(data = my_df, date_col = "bad_dates",
-                  group_col = "data_enterer", return = "dataframe")
+supportR::date_format_guess(data = my_df, date_col = "bad_dates",
+                            group_col = "data_enterer", return = "dataframe")
 
 # If preferred, do it without groups and return a vector
-date_format_guess(data = my_df, date_col = "bad_dates",
-                  groups = FALSE, return = "vector")
+supportR::date_format_guess(data = my_df, date_col = "bad_dates",
+                            groups = FALSE, return = "vector")
 
-## ----theme_lyon, out.width = "50%"--------------------------------------------
+## ----theme_lyon, message = F, warning = F, fig.width = 5----------------------
 # Load ggplot2
 library(ggplot2)
 
-# Create a plot
+# Create a plot and allow default ggplot themeing to be added
+ggplot(penguins, aes(x = species, y = body_mass_g, fill = species)) +
+  geom_boxplot(outlier.shape = 24)
+
+# Compare with the same plot with my theme
 ggplot(penguins, aes(x = species, y = body_mass_g, fill = species)) +
   geom_boxplot(outlier.shape = 24) +
-  theme_lyon()
+  supportR::theme_lyon()
 
-## ----nms_ord, message = F, warning = F, results = 'hide', out.width="50%"-----
+## ----nms_ord, message = F, warning = F, results = 'hide', fig.width = 5, fig.height = 5----
 # Load data from the `vegan` package
 utils::data("varespec", package = 'vegan')
 resp <- varespec
@@ -92,36 +127,33 @@ mds <- vegan::metaMDS(data[-1], autotransform = FALSE,
                       expand = FALSE, k = 2, try = 10)
 
 # With the scaled object and original dataframe we can use this function
-nms_ord(mod = mds, groupcol = data$factor_4lvl,
-        title = '4-Level NMS', leg_pos = 'topright',
-        leg_cont = c('1', '2', '3', '4'))
+supportR::nms_ord(mod = mds, groupcol = data$factor_4lvl,
+                  title = '4-Level NMS', leg_pos = 'topright',
+                  leg_cont = c('1', '2', '3', '4'))
 
-## ----array_melt---------------------------------------------------------------
-# Make data to fill the array
-vec1 <- c(5, 9, 3)
-vec2 <- c(10:15)
+## ----github_ls, eval = F------------------------------------------------------
+#  # List all files in a GitHub repository
+#  supportR::github_ls(repo = "https://github.com/njlyon0/supportR",
+#                      recursive = TRUE, quiet = FALSE)
+#  
+#  # Or list files in only a particular folder
+#  supportR::github_ls(repo = "https://github.com/njlyon0/supportR", folder = "R",
+#                      recursive = FALSE, quiet = TRUE)
 
-# Create dimension names (x = col, y = row, z = which matrix)
-x_vals <- c("Col_1","Col_2","Col_3")
-y_vals <- c("Row_1","Row_2","Row_3")
-z_vals <- c("Mat_1","Mat_2")
-
-# Make an array from these components
-g <- array(data = c(vec1, vec2), dim = c(3, 3, 2),
-           dimnames = list(x_vals, y_vals, z_vals))
-
-# "Melt" the array into a dataframe
-array_melt(array = g)
+## ----github_tree, eval = F----------------------------------------------------
+#  # Create a file tree diagram of a GitHub repository
+#  supportR::github_tree(repo = repo = "https://github.com/njlyon0/supportR",
+#                        exclude = c("docs", "man", ".github"), quiet = FALSE)
 
 ## ----rmd_export, eval = F-----------------------------------------------------
 #  # Authorize R to interact with GoogleDrive
 #  googledrive::drive_auth()
 #  
 #  # Use `rmd_export()` to knit and export an .Rmd file
-#  rmd_export(rmd = "my_markdown.Rmd",
-#             in_path = file.path("Folder in my WD with the .Rmd named in `rmd`"),
-#             out_path = file.path("Folder in my WD to save the knit file to"),
-#             out_name = "desired name for output",
-#             out_type = "html",
-#             drive_link = "<Full Google Drive link>")
+#  supportR::rmd_export(rmd = "my_markdown.Rmd",
+#                       in_path = file.path("Folder in my WD with the .Rmd named in `rmd`"),
+#                       out_path = file.path("Folder in my WD to save the knit file to"),
+#                       out_name = "desired name for output",
+#                       out_type = "html",
+#                       drive_link = "<Full Google Drive link>")