+++ NEWS.md
@@ -1,3 +1,18 @@
+# supportR Version 1.1.0.900
+
+Development version of `supportR`. Changes from preceding version will be listed here as they are made.
+
+# supportR Version 1.1.0
+
+Changes from preceding version are listed below
+
+- New function: `github_ls`. Lists contents of specified GitHub repository either recursively or only top-level/specified folder
+- New function: `github_tree`. Creates a file tree diagram for a specified GitHub repository
+- Refined and clarified the package vignette
+- `pcoa_ord` and `nms_ord` now include arguments for changing point size (`pt_size`) and opacity (`pt_alpha`; i.e., transparency). Changes to point size are reflected in the legend but changes to opacity are not reflected in the legend points
+- Fixed a typo in the message returned by `diff_check`
+- Added example in documentation for `crop_tri`
+
 # supportR Version 1.0.0
 
 This is the first fully-functioning version of the package. It currently has no ERRORs, WARNINGs, or NOTEs from `devtools::check`